+++ src/app/services/change-step.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
 import { BehaviorSubject } from 'rxjs';
-import * as _ from 'lodash';
 import { ISteps } from '../types/ISteps';
 import { PassengerService } from './passenger.service';
+import { passengerFormsAreValid } from '../shared/validators/passengerFormValidator';
 
 @Injectable({
   providedIn: 'root',
@@ -17,27 +17,22 @@ export class ChangeStepService {
   public continueButtonStatus$ = new BehaviorSubject<boolean>(false);
 
   public constructor(public passengerService: PassengerService) {
-    localStorage.setItem('passengersContact','');
-    localStorage.setItem('passengersInfo', '');
+    if (localStorage.getItem('passengersContact') &&
+    localStorage.getItem('passengersInfo')) {
+      this.changeButtonStatus(false);
+    } else if (!(localStorage.getItem('passengersContact') &&
+    localStorage.getItem('passengersInfo'))) {
+      this.changeButtonStatus(true);
+    }                   
     this.progressCondition$.subscribe((condition) => {
       const passengerServiceId = this.passengerService.passengers.subscribe((passengers) => {
         if (condition.passengers === 'active') {
-          const passClone = _.cloneDeep(passengers);
           this.passengerService.passengerContacts.subscribe((contacts) => {
-            let passengerFormsStatus = false;
-            if (Object.values(passClone).filter((el) => el).length) {
-              passengerFormsStatus = Object.values(passClone)
-                .filter((el) => el)
-                .flat(Infinity)
-                .map((pasArr) => pasArr.formIsValid)
-                .every((isValid) => isValid);  
-            }
-            const passengerContactFormStatus = contacts.formIsValid;
-            if (passengerFormsStatus && passengerContactFormStatus) {
+            if (passengerFormsAreValid(passengers, contacts)) {
               localStorage.setItem('passengersContact', JSON.stringify(contacts));
               localStorage.setItem('passengersInfo', JSON.stringify(passengers));
               this.changeButtonStatus(false);
-            } else if (!(passengerFormsStatus && passengerContactFormStatus)) {
+            } else if (!(passengerFormsAreValid(passengers, contacts))) {
               localStorage.setItem('passengersContact','');
               localStorage.setItem('passengersInfo', '');
               this.changeButtonStatus(true);
